Switch hh_server and HHVM to use Rust parser errors (with opt-out flags)
Summary:
This is fully ready, so this diff not only adds the flags, but also enables the whole thing (so the flags are only in case something bad happens, and hopefully never even will get used).

Fred: is this fine from Open Source perspective? I don't expect trouble, but this time there **were** some non-trivial changes around namespace handling (because all the "names in scope" logic is heavily using immutable maps with cheap backtracking, that we don't have in Rust). We could mention the flag in release, and if there are unexpected problems they can enable it locally and let us know what needs fixing.

Reviewed By: shiqicao

Differential Revision: D17496268

fbshipit-source-id: 0d05ca6e26337b0ae68f0a8cda3de215f6f1b337