Add IterTypeOp immediate to LIter ops
Summary:
If we know that the base is unchanged, we can make two significant optimizations to specialized iterators:
- We can bump the refcount lower bound on the key and value during iteration; they're held by the constant base. Doing so will eliminate almost all refcounting ops.
- We can iterate with a pointer rather than a base an index. Doing so will eliminate ~2 registers that we currently hold, since we currently need a register for the base, a register for the offset, and (for packed arrays) a register for the shifted index.

The base is always constant for non-local iterators. It's also constant for *most* local iterators; the analysis we do captures "safe" updates which set the value for the current iteration key, but for most local iterators, there are just updates at all. We currently drop this analysis after doing it in HHBBC; in this diff, we pass it along to irgen. We provide the information in a new "IterTypeOp" immediate for the LIter ops. The IterTypeOp also includes a "NonLocal" type which is implicitly the type of the Iter ops; adding it here makes the next diff in this stack cleaner.

Reviewed By: jjgriego

Differential Revision: D17399419

fbshipit-source-id: 6aa7d9a7c953fdb060402734ef1484d5a877fbe8