Support pointer iteration for iterators with constant bases
Summary:
The "local iterator" optimization lets us avoid COW in cases where we update an array's current iteration key, but it also prevents us from doing the pointer-based iteration that we would typically use to iterate over an array in C++. This wouldn't be a problem if it was specific to cases where the array is mutated, but we also use local iteration for cases where the base isn't touched at all. As a result, about half of all iterators (by weight) are local and may include an update to their base changed during iteration. (The weight of an iterator is the number of times that particular loop gets executed. The contents of the loop don't matter for iterator performance.)

In D17399419, we introduced a new IterType bytecode immediate that distinguishes between LocalBaseConst and LocalBaseMutable. In this diff, we use pointer-based iteration for the types LocalBaseConst and NonLocal. Together, these two types cover almost all iterators (by weight).

In addition to excluding LocalBaseMutable for correctness, we further restrict when we use pointer-based iteration:
- For PackedArrays, we only use it for value-only iteration. (For key-value iteration we need to materialize the index anyway, but value-only iteration over PackedArrays is quite common.)
- For MixedArrays, we only use it when the array doesn't contain tombstones, which is usually the case.

This diff is a small instructions and CPU win on its own, but the real benefit is that when we specialize code for an iterator, we can now keep the iteration state in a single register (the current element pointer) as opposed to three (the base, the position, and the "shifted position"). The "shifted position" is only used when we're loading an element from the array, but given that loops already introduce quite a bit of register pressure, it still causes spills in lots of examples that I've looked it. We'll shave 1-2 arithmetic ops from each loop iteration, plus potentially quite few a more ops saved from avoiding spills.

Reviewed By: jjgriego

Differential Revision: D17470928

fbshipit-source-id: c7d24001897e19049783e0e5b6008d00f8bbaa30