Consolidate timeout surprise flags into a separate enum
Summary:
The goal is to allow introduction a third kind of timeout - a "pre-timeout" callback into PHP, while clearing up a surprise flag at the same time.
We can offset the kind of the timeout into its own flags enum on the RID and use a single Timeout surprise flag to represent all kinds.

Reviewed By: alexeyt

Differential Revision: D17454729

fbshipit-source-id: 9834bd634d555bcaf563e8398c4cd308a10916e0