Fix race in jit serialize server shutdown
Summary:
If a server is running in serialize and exit mode, it may crash if the
server is shutdown at roughly the same time as retranslate all goes to trigger
server shutdown.  This is because the server pointer is reset without
synchronization or waiting between the retranslate all thread, and the main
thread that resets the pointer.

Reviewed By: binliu19

Differential Revision: D17182441

fbshipit-source-id: 9c8d4341d16f77e84130cb31e97bc51fe4882aeb