Split prologue handling enterTC() logic to a separate code path
Summary:
There are two main ways how to enter TC: at a regular translation resolved from
SrcKey (or a resumeHelper that can resolve it), or at a prologue.

Split the enterTC code path and specialize for each use case. Kills a totally
unnecessary branch at each TC entry alongside hacky workaround for a branch
predictor misses and reduces the amount of passed arguments from 6 to 3 or 5
registers.

Makes it also possible to change the way how we call prologues.

Reviewed By: ricklavoie

Differential Revision: D17536651

fbshipit-source-id: c1cdff8168fb474bf48a708049a1760b66b14aa5