Respond to client even when an unexpected exception is thrown
Summary: Certain code paths weren't expected to throw an exception, but did anyways (such as `make_context_from_document_location`). When the client issued a request, we did actually catch exceptions from these code paths, but we failed to send an error response back to the client. (Normal IDE code paths would have an inner exception handler and return the `Error` object to be sent back to the client directly, without relying on exceptions to do so).

Reviewed By: ljw1004

Differential Revision: D17541786

fbshipit-source-id: 73acd3a6fdb3deeb8b8e7fb9c2d679f93237c75d