Remove instances of primitive aliasing
Summary:
The compiler transparently aliases some primitive types to canonical ones. This has been a typechecker error for a long time, but there are a few instances in systemlib and extensions. This codemods the aliases to the canonical versions in preparation for removing the aliasing behavior.

This is a step in removing behavior differences between the typechecker and compiler so that we may pull the namespace elaboration code earlier in the pipeline.

Reviewed By: fredemmott, alexeyt

Differential Revision: D17533445

fbshipit-source-id: c513e966aef7e445a7db43867ea01ececf731be9