Fix dependency collection for enum constants
Summary:
When we encounter a constant whose type is an enum type, we add a dependency on this enum's value so that we can generate an initializer for this constant. This commit changes this behaviour in several cases:
1. Add a check: the class type is an enum type if it's not a builtin type AND it's not a typedef.
2. If this constant is one of the enum's values, we don't need an enum value to initialize it (consider
```enum MyEnum: int {
  FIRST = 1;
}
```: the type of `MyEnum::FIRST` is `MyEnum`, but we initialize it with an integer, not with one of `MyEnum`'s values).
3. The constant's type can include another class's type. Sometimes we can be sure it's an enum and add a dependency on one of its values. However, it doesn't have to be an enum if it's a type parameter (e.g. `Vector<ArbitraryClass>` is still compile-time-initializable).

Reviewed By: CatherineGasnier

Differential Revision: D16940131

fbshipit-source-id: 23a0fd7dab477045aa95e6fb4c9a60df73ef7cdd