Fix docblock for PHP\var_dump
Summary:
Some users reported that docblocks for builtins like `PHP\var_dump()` were not appearing.  Instead, users got the message "This symbol was added recently, please rebase".

The fix for this was to use the fullname of the function, which included the expanded namespace `HH\Lib\PHP\var_dump()`.

The original intention of the warning message about rebasing was the user might be connected to a saved-state that was not in sync with the user's files on disk, and that the warning message would encourage them to resync. Let's rewrite the message as well.

Reviewed By: 2BitSalute

Differential Revision: D17563001

fbshipit-source-id: 3d8cb948ada92635d05279e7f97d33206a7404be