Reset naming path after saving
Summary:
Every time we upload a naming table for remote type checking, we save it into a temp folder and then clean it up when we're done type checking. Because of this, any operation requiring the naming table after the remote type check resulted in a crash.

Before this change (Julia is the author - feel free to commandeer), we would replace the current DB with the one we just saved. Instead, we should save then point ourselves back to the original DB.

Reviewed By: arxanas

Differential Revision: D17538412

fbshipit-source-id: 4090bf98c5d84298747694ef577633e48f323422