Make classname/typename real aliases
Summary:
These were magically aliased to primitive types in the compiler, but they were defined as type aliases in an HHI file. In this diff we make them real type aliases (by copying them directly from the HHI) and remove the magic aliasing.

This contributes to the unification of autoimport behavior between the typechecker and compiler, and eventually to the elimination of HHIs themselves.

Reviewed By: oulgen

Differential Revision: D17538644

fbshipit-source-id: eba3eb844e5c1598ca7d0b2838b2ea21f16fae2b