Move ActRec::Flags::MultiReturn to CallFlags::Flags::HasInOut
Summary:
Do not pass a flag whether we are calling an inout method via ActRec. Instead,
use the CallFlags structure.

Add a plumbing for CallFlags in doFCall() and fcallUnpackHelper().

This flag will be eventually killed altogether once inout enforcement moves to
the infrastructure used by references and support for reference arguments is
removed, but it is needed for now.

Reviewed By: ricklavoie

Differential Revision: D17589964

fbshipit-source-id: 7c0a3d02c8b13ae95fe18373b8620d29292876e6