Allow for the use of disallow_array_literal outside of fb (#8576)
Summary:
I applied the typechecker setting `disallow_array_literal` and made my code use `darray[...]` and `varray[...]` instead of `[...]` and `array(...)`. HHVM's built-in hhi's don't typecheck with this setting turned on. The PR attempts to mitigate that.

As far as I understand the default values in hhi's are not used at runtime. I did check for PHP examples of each function, to make sure that the `d/v`array type was possible. I did not try to rule out if a function could also take the other type.
Pull Request resolved: https://github.com/facebook/hhvm/pull/8576

Differential Revision: D17603372

fbshipit-source-id: 149607c77be1166e54c2dfaba8c599865550d855