Add --ranked-autocomplete flag to hh_client
Summary:
This flag is used to turn on/off the ranked autocompletion functionality for Hack.

This flag will be passed in if the user passes the `nuclide_ranked_hack_autocomplete` GK in Nuclide. I want to add the flag first before adding the GK logic to Nuclide. Refer to the next diff for more implementation details on the client side.

This diff does not contain any behavioral changes. The concrete implementation for the ranked hack-autocomplete will come later.

Differential Revision: D17595699

fbshipit-source-id: f59d37fd96bd8d1b6d2a9fbb447611da6f0d5db4