When adding a new dependency, make sure we add its signature dependencies
Summary: Dependency collection happens in two steps: first we typecheck the function or the class of the method we want to extract, then we fetch signature dependencies of the dependencies we collected during the first step. However, the dependencies in the second step might have their own signature dependencies, e.g. ancestors for classes. In this commit we call `add_signature_dependencies` every time we add a new dependency to the set of dependencies.

Reviewed By: CatherineGasnier

Differential Revision: D16984655

fbshipit-source-id: 3298609503c79fdd0ccd02a6111d4ed15962dd95