Implement ProfileDecRef IR op
Summary: This opcode is useful when we call out to native helpers during profiling translations but emit inlined code during optimized translations. We'll use this opcode to dec-ref the iterator key and value during profiling IterNext, because it's important to emit good code for dec-refs for specialized iterators. (More so than inc-refs.)

Reviewed By: jjgriego

Differential Revision: D17242284

fbshipit-source-id: 815b547567652245085b7b88960d16fd6e0e33f1