Encapsulate MixedArray key-types tracking
Summary:
MixedArray keyTypes is a bitset of four bits tracking keys in the MixedArray. The four key types are int, static str, non-static str, and tombstone. ("tombstone" is not really a key, but it can appear in a key position, and knowing the array is tombstone free is very useful e.g. for iterators or for lookups.)

In the original version of this diff, I left a lot of bit-wise arithmetic in different places that did logic on the bitset. In this diff, I collect it all in one place and hide it behind a "MixedArrayKeys" abstraction. MixedArrayKeys has quite a few methods, because we want to make sure that we can do every type of bulk update to the bitset efficiently, but it does hide the arithmetic and in particular the key bit constants from users. MixedArrayKeys has roughly four types of method:

- "getters" that check that all keys match a condition, like "all uncounted". Getters are pessimistic, as we can only guarantee the absence of a key type, but that's good enough for optimization.
- "bulk setters" that implement the update for things like "compact", "renumber", etc.
- "point setters" that record a particular insertion or deletion.
- debugging helpers.

Reviewed By: alexeyt

Differential Revision: D17523255

fbshipit-source-id: 75e78d688cf9d9894d24cba9586479af014507ab