If a class implements a built-in interface, generate stubs for the implementations of the interface's methods.
Summary: Normally we only generate stubs for methods which we use. However, if a class implements a built-in interface (e.g. Stringuish), we have to provide implementations for this interface's methods (otherwise we get an error "this class doesn't implement...") even if we don't use those methods in the function or method we are extracting.

Reviewed By: CatherineGasnier

Differential Revision: D16984953

fbshipit-source-id: bcac274550b519155924bb4852562ca2e3160f7f