Make the helper function which "returns" value of any type return `nothing` instead of `T` (so the function is no longer generic).
Summary: When extracting standalone code, we sometimes need to "produce" values of an arbitrary type, e.g. to initialize properties or default values for function parameters. We achieve this by calling `extract_standalone_make_default<T>()`. This function used to be generic, but this commit makes it return `nothing`, so that the type parameter is no longer needed. This will allow to use the function when extracting code from `// partial` files.

Reviewed By: CatherineGasnier

Differential Revision: D16985051

fbshipit-source-id: f33e3a94faf1ff6683af81dae1fd00a324796c4a