Favor standard elaborated identifier in more cases
Summary:
The boolean returned from `elaborate_id_impl` is misnamed based on how it's used. It's now used by Hhbc_id to determine if the identifier is "finished", i.e. whether we should fall back to compiler-specific transforms defined in Hh_autoimport.

Previously, we'd return true only if the identifier was renamed via an explicitly-written `use` statement. But, an identifier is finished if we HH-autoimport it, if we namespace alias it, and if it's a special global name! This diff updates `elaborate_id_impl` accordingly.

This should not cause a behavior change:
- The compiler doesn't autoimport functions or constants.
- Namespace aliasing requires a namespaced name, and everything in Hh_autoimport is a bare name.
- The list of HH-autoimported types is the same, so if it's caught and transformed earlier, it's inconsequential.

This change will allow us to begin removing types from Hh_autoimport, eventually unifying the implementation between the typechecker and the compiler.

Differential Revision: D17509802

fbshipit-source-id: baa04bfa33dd25a5d6f7f147ab21b5734efe7a04