Remove GenerateFFJavaScript and /js/ files using it
Summary:
At some point we'll like to change to generator from OCaml to Rust, so it's time to start cutting out the fat. As far as I know those were some early prototypes of FFP, and are unused for ages.

There is still `/js/full_fidelity_schema.json` which is exposed through `hh_parse --schema` and `hh_client --full-fidelity-schema`, which in turn is referenced from `hh_parse_service`, which seems to be used, so leaving this one untouched.

Reviewed By: kmeht

Differential Revision: D17672515

fbshipit-source-id: da8ae76f13312668dfba459864f557ef55b408da