Simplify ctx type refinement when inlining
Summary:
Note that beginInlining()'s ctxType argument is always passed as ctx->type(),
so nuke it and then simplify various branches of ctx refinement.

Note that SpillFrame is now issued in the same function, so replace LdARCtx
opcodes with the same ctx that goes into SpillFrame.

Then since ctx type refinement no longer depends on SpillFrame, move it above
the SpillFrame in preparation for the next diff that will merge it into
DefInlineFP.

Also, do not fail on assertions if the refined context is TBottom. Fixes some
outstanding crashes in debug mode.

Plus fix a bug where we were conflating context with prologue context for
closures when computing the inlining cost.

Reviewed By: ricklavoie

Differential Revision: D17643922

fbshipit-source-id: d67438088aa6efd19bbffd29cb8a646e8cf1c3de