Fall back to userland autoloading when we can't update the map
Summary:
Currently, native autoloading is an all-or-nothing business. If we can't update our native autoload map, we say we're proceeding with stale data (and we may not have *any* data) and try to continue with native autoloading.

We're unable to run the native autoloader on `/var/www` right now. It's owned by root, and Watchman can't report on it unless we have write access to it. The long term solution is still to find a way to autoload `/var/www`, but this diff would allow us to load a user's WWW repo natively while allowing `/var/www` to function as it does today.

Reviewed By: billf

Differential Revision: D17618650

fbshipit-source-id: 86901f8db5863a7de7b3f1ffa6b56a06a6991a70