Improve error tracking during global inference
Summary:
Hh wasn't design to show all typing errors to the user, but only some of them. For global inference we want to all of them in order to track which type variable wasn't solved correctly and avoid adding an incorrect annotation.
To do this, we were using error handle. However, because of the previous explanation, this wasn't good enough: if a type yielded an error was reused as a lower bound (or an upper bound) in an other type, then we could still be solving this type b without seeing an error.
This was due to the process of binding a type variable immediatly if its lower bound and upper bound were the same. Thus, we deactivate this optimization when using global inference.

Reviewed By: CatherineGasnier

Differential Revision: D17668069

fbshipit-source-id: dca17f27f422efccbb6b78781bc15ab9f7ad6d88