Extend the solved constraint graph with the position of type variables
Summary: When we are solving the global constraint graph we are binding each type variable to a value. This operation removes the type variable from the `tvenv` entry which has the side effect of removing the position of the type variable from the `env`. As we will need this position later on to rewrite files we need to store it somewhere else.

Reviewed By: CatherineGasnier

Differential Revision: D17423788

fbshipit-source-id: 37a56efeb5535c2c6f25d1ad44a897c7d63bba8d