Add a new submode to global inference that adds explicit type annotations
Summary:
The last missing piece of global inference was a way to add the type annotations we were able to infer back in the code. This new submode does exactly that.

To use it execute (once the other steps of global inference are executed and you have a solved the global constraint):
```
hh --global-inference rewrite [solvedenv]
```

For now a type is rewritten if it passes the following two filters:
- No type variable in the type triggered an error during the merging phase and solving phase
- It is deemed "acceptable" (meaning it is denotable and contains no mixed/null type)

Reviewed By: CatherineGasnier

Differential Revision: D17559169

fbshipit-source-id: 2b41ffc8eac0793e3eba021f6d452d0e306d2299