Add tests for eager mode and correct graph construction
Summary:
This diffs is still ongoing work.
It is a draft to add tests to global inference. However, due to the way global inference currently works, theses tests are not working (I still got an hardcoded hard drive location in global inference which prohibits from running the tests in parallel).

Reviewed By: CatherineGasnier

Differential Revision: D17477406

fbshipit-source-id: 289b7f23ade72c3112c665648f56ae0187eefc90