Move the region formation workaround for ignoring inferred return type deeper
Summary:
Do not just callUnknown() when we are forming region, as that approach no
longer works with the other diff that moves callee of the Call and CallUnpack
instructions from their extra data to their source.

Instead, pass a flag indicating whether we are forming region and use that to
disable inferrence.

Reviewed By: markw65

Differential Revision: D17683654

fbshipit-source-id: 3a1a9566546d24cee7579145fdc312a31fb7831d