Raise an error when trying to interpolate a variable that doesn't exist
Summary:
In LSP tests, we use variable interpolation for requests and responses so that we can produce semantically-correct suggestions. For example, this allows us to suggest that the user replace their failing response with a payload containing `${php_file_uri}` as one of the values, instead of a temporary path like `/tmp/aBcDeF/foo.php`, as that temporary path won't be correct for the next run and it'll be harder to update the test to be correct.

We currently don't warn if the developer has written a variable interpolation expression for a variable which doesn't exist. This means that any variables in the payload will go silently un-interpolated, which may result in incorrect or confusing behavior.

Raise an exception if we find an interpolated variable in the payload that was not defined in the variable map.

Differential Revision: D17679505

fbshipit-source-id: ade89cee91e799997751fa8f388bb77c4a4ed8a6