Rule-of-three for classes owning Transport*
Summary:
This diff serves as a lighter-weight alternative to D17676294.

The "rule of three" essentially says that if a class or struct declares a custom destructor, it should almost never use the default copy constructor or assignment operator. Here we see the rule in action. If any of these classes are default copy-constructed, we'll dumbly copy the pointer data member without calling incRef. Both the copy and the original will run their destructor, and we'll get two decRefs per incRef.

This diff deletes the copy constructor to make it clear to the programmer that copying is not safe. We could also implement a simple safe copy constructor which calls incRef in its body, but I think it's simpler to forbid something than to write an unused code path.

Reviewed By: markw65

Differential Revision: D17688827

fbshipit-source-id: 56e0ded3db4f8576a20a1d9c3b704f90dd97ce61