Skeleton of hh_decl and other related command-line tools
Summary:
This commit stubs out the `hh_decl` and `hh` executables per the architectural diagram. For `hh_decl` in particular, this commit implements the IPC mechanisms necessary for prototyping the various components.

The `Service` abstraction in particular is reusable for other components, if there are other Rust executables that intend to run services on unix domain sockets.

Reviewed By: 2BitSalute

Differential Revision: D17242391

fbshipit-source-id: 270e4bb0fb82835880e0c2403f061ea4a67c4efa