Delete elaborate_toplevel_defs_nast
Summary:
We don't use this anywhere. To prevent the two copied code paths from diverging, delete the unused one.

Note: Rather than switching the one that operates on the ast with the one that operates on the aast, I keep the ast one. This is because we might want to integrate the ast pass directly into lowering. And the nast one has the risk of getting fragmented as our ast to aast conversion is spread out all over the place and not centralized.

Reviewed By: kmeht

Differential Revision: D17695712

fbshipit-source-id: cdf4a02edbf17e5df7a03f267e1d8a2d2207422f