Revert visibility of class/type constants.
Summary:
This was partly built but abandoned as a direction we don't intend to pursue further. It also blocks me from using the presence of visibility modifiers to differentiate constant declarations from const property declarations.

Revert the stack from D16043672 to D16087542, D15981533, and D16137608 except for some useful refactors:
* refactors in hphp/hack/src/typing/typing_extends.ml from D16081937 and D16085948
* the bits of D16075262 that were plumbing type const modifiers (in general, as opposed to visibility ones specifically) into the ast
* the parts of D16069842 that changed class_const from a tuple to a struct

Reviewed By: dabek

Differential Revision: D17680976

fbshipit-source-id: 2342037619ddfaeb8ee1414e8871af5c77fb5eaa