Switch coercion to treating dynamic as a bottom type
Summary:
Currently when determining whether or not a coercion is valid, we have custom logic in Typing_coercion that matches against some cases of a type. This has proven to be less than ideal in practice, especially as we think about more complex coercion logics.

One thing that is clear is when coercing from a like type to an enforced type we generally want to ignore the dynamic portion of the type. For example:

For `~T ~> E` we want the coercion to exist only if `T <: E`.

The one case where this treatment of dynamic isn't desirable is with reified generics. In that case a `C<T>` is enforceable but we don't want to ignore the dynamic in the type `C<~T>` when deciding `C<~T> ~> C<T>`.

To implement this I thread through a flag in typing_subtype which is most of this diff.

Reviewed By: CatherineGasnier

Differential Revision: D17682757

fbshipit-source-id: f67eeedac33325b15ed67201b4fed2f2165562b1