Implement complex coercion by treating dynamic as bottom
Summary:
This builds on {D17682757} by adding coercions to like types. For a type `C<T>` we want to allow coercions the coercion `C<~T> ~> ~C<T>`. Note that it is always sound to coerce in this case because we can choose to coerce to dynamic which is part of the union. We do not want to allows do this however since it may be surprising to developers.

By treating dynamic as the bottom type we can test if the given source type is a subtype of the target type ignoring any untrusted parts since they will be "lifted" to the outer dynamic.

Reviewed By: CatherineGasnier

Differential Revision: D17691430

fbshipit-source-id: 6e5a8b95e95a9f6c830e9d05ae7ae520ce4c0d53