Add more knowledge of rematerialization to vasm-graph-color spilling heuristic
Summary:
Up until now the spilling heuristic in vasm-graph-color was
agnostic as to whether a Vreg could be rematerialized or not. However
this information is important because you generally want to spill a
Vreg which can be rematerialized over one which cannot (because
rematerializing it is so much cheaper). This is difficult to calculate
because precisely what Vreg can be rematerialized or not at any given
program point is a function of which Vregs are currently live and in
physical registers at that point. Since the spilling heuristic is
looking forward into the program, it does not know this information
(and cannot, since it depends on its own decisions).

To help work around this issues, introduce the notion of "trivially
rematerializable". This is a Vreg which is guaranteed to always be
rematerializable (because it's defining instruction has no
sources). This, right now, is just the ldimm instructions, which is
fine because those are the majority of rematerialization cases
anyways.

Modify the spilling heuristic to take into account whether a Vreg is
trivially rematerializable or not. Apply a scale to the spill weight
depending if it's not rematerializable or not. This makes
rematerializable Vregs more attractive to spill as opposed to
non-rematerializable ones. It's not correct to simply always prefer
rematerializable instructions. For example, it might be better overall
to spill a non-rematerializable Vreg across a loop, if it means we can
keep an immediate in a register inside the loop. The weight tries to
capture this dynamic.

Rematerialization can be leveraged in other ways as well. If we know a
Vreg is rematerializable as an immedate, when spilling it, we can
instead write the immediate directly to the spill slot. This cuts down
on Vreg lifetimes. We can go even further. If a Vreg is spilled and
then copied, and we know that we won't be able to coalesce the copy
away (because their lifetimes overlap), we can leverage
rematerialiation. If the copied Vreg is rematerializable, instead of
the memory to memory copy, we can instead just write the immediate to
the copy's destination. It might also be profitable to just load the
value into a register. Because of this, phijmps and copies can now
trigger reloads.

Phis now have the notion of a "triviality boundary". This is when the
input to the phi is trivially rematerializable, but the output is
not. If the input is spilled, we won't be able to kill the spill slot
because any usages after the phi won't be rematerializable. In this
case, we want to materialize the immediate directly into the spill
slot right at the phi. This will kill the spill slot up until that
point.

Finally make some modifications to the "find defining instruction"
logic. Its tracking of the visited set wasn't great, meaning it could
explore an exponential number of paths in some degenerate cases
(taking minutes). Resolve this by caching intermediate resolutions on
the way. However this doesn't quite work with recursive definitions,
so add special handling to those.

Reviewed By: mxw

Differential Revision: D17599448

fbshipit-source-id: 64d60b7d06143367b542f55ffd89bd6ae3c43231