Hoist spills out of loops in vasm-graph-color
Summary:
If a Vreg is in a physical register at the entry of a loop, is not
used within the loop, but is spilled somewhere within the loop, it's
particularily bad. This is because the Vreg must be reloaded on the
back-edges of the loop, as well as spilled within it, every trip
though the loop. This situation must be avoided at all costs.

The spiller has some heuristics to avoid this, but they're not super
effective. They're a hold over from an earlier version of the spilling
heuristic, and does not work well with rematerialization. Get rid of
these and try something new.

Rather than try to detect this situation ahead of time, detect it
after the fact and improve things. After the spiller runs, but before
we resolve block mismatches, run a "loop spilling hoisting" pass. This
looks for situations as described above, and attempts to hoist the
Vreg's spill(s) out of the loop to the predecessors of the loop entry
block. It only does so when it calculates that it is profitable.

Tweak the spilling heuristic to try to guess when this behavior will
happen to bias it towards spilling things in loops which we know we
will eventually hoist.

With this, we can remove the special block initialization logic for
loops. We can also simplify the metadata we need to calculate for
loops. We no longer need to calculate maximum register pressure within
loops (which is expensive). Instead we only have to calculate Vreg
usages within the loop (which is simpler).

Reviewed By: mofarrell

Differential Revision: D17599507

fbshipit-source-id: c7ff023f21926a414d059d134dd7c62ec9d728bd