Make autoimported types case-sensitive
Summary:
Autoimported functions and constants are already case-sensitive. This makes types operate the same way (with some notable exceptions, i.e. the primitive types).

The reason this diff is so small is that this is the first/only behavior change in this stack of diffs, which I wanted to isolate.

Differential Revision: D17620923

fbshipit-source-id: 2c88ada89984690ef705e963c4ddf3a305de3dc4