Fix hhbbc to attribute addelemc const-propp'd dicts correctly
Summary:
We weren't marking the empty array at the start of the chain, so do
that. Also tweaks `provTagHere` to grab the _last_ line of the range instead of
the first--we may want to revisit that if we decide we really want arrprov
marked at the start of an array literal.

Reviewed By: mxw

Differential Revision: D17467778

fbshipit-source-id: 9bd6d672ac5a3a6c971fb77452d1fd1a55268dbb