Derive eq for shallow decls, implement eq for decl_ty
Summary: If we are to store decls off of the OCaml heap (via ocamlrep in cachelib), we cannot use polymorphic equals, since it will not compare on-heap and off-heap values correctly. We should derive monomorphic equality functions and use them instead. This diff is a first step on that path.

Reviewed By: arxanas

Differential Revision: D17638283

fbshipit-source-id: 5c7eddd218f194e6842c7f9c1873013f62676371