set use_numa to false on single-socket hosts
Summary:
We already try to set RuntimeOption::EvalEnableNuma to false on single-socket systems, so it doesn't change the behavior.

Also move enable_numa() to numa.h/cpp.

It won't be necessary to check environmental variable "HHVM_DISABLE_NUMA" in multiple places.  The check in init_numa() should be sufficient, and other places can look at numa_node_mask or numa_num_nodes, which will both be 1 when the environmental variable is present.

Reviewed By: markw65

Differential Revision: D16167010

fbshipit-source-id: 4bd22cb57363a94e53e6ed82d3bb4c9481f97732