Allow unset on when indexing a dynamic value
Summary:
This adds support for `unset($dyn[$idx])` when `$dyn: dynamic`. This will also allow like types to work here as well. Since `dynamic` is now a valid type in this case, I added a way to ignore printing dynamic in the error string that we report as well.

This cleans up a few more error cases.

Reviewed By: vassilmladenov

Differential Revision: D17726306

fbshipit-source-id: f4918c2e6251eb769e4ba65e5a14ed08699411b9