Add lvalSilent() and lvalForce() to array API
Summary:
lvalSilent() is a non-throwing lval() which returns a falsey array_lval
if the key is not present.  This is added to the ArrayData vtable.

lvalForce() implements the legacy lval() behavior, where null is written
when the key is not present.  It is implemented on the Array wrapper
only.  This is slightly less efficient.

Reviewed By: jjgriego

Differential Revision: D16826251

fbshipit-source-id: e7039d97e4a85ab1a5ff083f953a34e51b298528