Generate declarations for the dependencies from the standard library
Summary: At the moment, functions whose names start with `HH\` are considered to be builtins and declarations for such functions are not generated.  This is not correct: e.g., the standard library functions live in the `HH\Lib` namespace (typically accessed using shorter namespace paths like `PHP` defined in the `auto_namespace_map`).

Reviewed By: andrewjkennedy

Differential Revision: D17710585

fbshipit-source-id: abf833794378dec37858875e489e12fe72bf70ef