Remove ft_deprecated and ft_decl_errors
Summary:
There are a number of fields in `fun_type` that don't really belong in a "type", but should rather be stored with the "member". One such is `ft_deprecated`, also the `ft_decl_errors` field that is used to redirect decl errors to the typing phase.

By moving these fields into a new `fun_elt` record, this gives us the opportunity to return a general type from `get_fun`, rather than a `fun_type`, which can be used in future to support intersections of function types, and to separate generics (`ft_tparams` etc) from function types (i.e. a proper "forall" type).

Error message position now more accurately points to the method identifier in the invocation, which explains why some expected output from tests needs to be changed.

Reviewed By: manzyuk

Differential Revision: D17685288

fbshipit-source-id: 3b0305d7f0daafe69180c97c784a83699ea4029a